pyproject.toml
[project]
name = "web-app"
version = "0.1.0"
description = "Flask web app"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.6",
    "requests>=2.32.3",
    "types-requests>=2.32.0.20241016",
    "flask-sqlalchemy>=3.1.1",
    "cotc-common",
]

[tool.uv.sources]
cotc-common = { path = "cotc-common" }
